Bin to Double, Random, Tests, Shelf Print
Bins changed to hold items of type double
Bin capacity changed to 1.0
Random item generation now works
Introduced a generalized runTest method
Added a few basic test cases
Added operator<< for Shelf objects.
Changed top comments to include all names
Changed top comments to only include descriptions in .h files

+++ BinPacking/BinPacking/Bin.cpp
@@ -1,8 +1,7 @@
-// Bin.cpp
-// Contains a specific amount of things.
 // Meghan Haukaas
 // Chris Ruiz
 // Stephen Belden
+
 // 2015-Oct-25
 
 #pragma once
@@ -14,15 +13,15 @@ using std::endl;
 
 Bin::Bin() : filled(0) {}
 
-unsigned int Bin::getFilled() const{
+double Bin::getFilled() const{
 	return filled;
 }
 
-unsigned int Bin::getRoomLeft() const {
+double Bin::getRoomLeft() const {
 	return capacity - filled;
 }
 
-void Bin::insert(unsigned int object) {
+void Bin::insert(double object) {
 	if(object > getRoomLeft()) {
 		cout << "FATAL: Tried to insert an object into a bin with insufficent room."
 			<< endl;
